--- lunespy/client/transactions/leasing/constants.py
@@ -1,3 +1,0 @@
-BYTE_TYPE_LEASING = b'\x08'
-INT_TYPE_LEASING = 8
-DEFAULT_LEASING_FEE = int(1e5) # 0,00100000 Lunes
